Initial check-in
Demo instrument cluster dashboard application to display instrument
cluster mock-up with speedometer, tachometer, temperature, and fuel
gauges.  A space is provided in the center of the dashboard for the
overlay of the receiver application's remote display.

Note that the gauges themselves and their animation are based on the
Qt dashboard example:

https://doc.qt.io/qt-5.11/qtquickextras-dashboard-example.html

The associated qml source files are under the BSD license, other files
are Apache 2.0.

Additionally, note that the application uses the windowmanager, but not
the homescreen service, as the demo instrument cluster does not use the
homescreen.  There is also an implicit assumption that the windowmanager
configuration allows the application to claim the homescreen role and
that the surface for that role will be 1920x1080 portrait mode.

Change-Id: I413fa165125813757ab4712993320440ed641f32
Signed-off-by: Scott Murray <scott.murray@konsulko.com>